requirements.txt
streamlit>=1.37.0
authlib>=1.3.0
requests>=2.32.0
psycopg[binary]>=3.2.1
SQLAlchemy>=2.0.32
alembic>=1.13.2
python-dotenv>=1.0.1
APScheduler>=3.10.4
pydantic>=2.8.2
pydantic-settings>=2.2.1
sendgrid>=6.11.0
pytz>=2024.1
plotly>=5.18.0
pandas>=2.1.0
